+++ CHANGELOG.md
@@ -18,6 +18,7 @@
     * Improved twig `authorize()` function to work better with nested rule parameters
     * Output the current username that Scheduler is using if crontab not setup
     * Translations: rename MODULAR to MODULE everywhere
+    * Optimized `Flex Pages` collection filtering
 1. [](#bugfix)
     * Regression: Fixed Grav update bug [#2722](https://github.com/getgrav/grav/issues/2722)